decimal无法json问题

+++ www/admin/views_statistics.py
@@ -127,7 +127,7 @@ def get_chart_data(request):
     register_cache = int(cache_str[0].strip()) 
     today_register_cache = int(cache_str[1].strip()) 
     today_order_cache = int(cache_str[2].strip()) 
-    today_balance_cache = int(cache_str[3].strip())
+    today_balance_cache = float(cache_str[3].strip())
 
     return HttpResponse(
         json.dumps({
@@ -137,7 +137,7 @@ def get_chart_data(request):
             'today_register_count_chart_data': [today_register_x_data, today_register_y_data],
             'today_order_count': sum(today_order_data.values()) + today_order_cache,
             'today_order_count_chart_data': [today_order_x_data, today_order_y_data],
-            'today_balance': sum(today_balance_data.values()) + today_balance_cache,
+            'today_balance': float(sum(today_balance_data.values())) + today_balance_cache,
             'today_balance_chart_data': [today_balance_x_data, today_balance_y_data]
         }),
         mimetype='application/json'
